+++ CoC.Frontend/Perks/History/Scholar.cs
@@ -7,14 +7,14 @@ namespace CoC.Frontend.Perks.History
 {
 	public sealed partial class Scholar : HistoryPerkBase
 	{
-		private float delta;
+		private double delta;
 		public Scholar() : base(ScholarStr, ScholarBtn, ScholarHint, ScholarDesc)
 		{
 		}
 
 		protected override void OnActivation()
 		{
-			float oldVal = baseModifiers.magicalSpellCost;
+			double oldVal = baseModifiers.magicalSpellCost;
 			baseModifiers.magicalSpellCost -= 0.2f;
 			delta = oldVal - baseModifiers.magicalSpellCost;
 		}
